Add updated package-lock.json file to commit
My solutions use objects to represent each station; these objects
are stored in 'stations.json' (for the base problem with the
original map) and in 'bonus-stations.json' (for the bonus problem
with the extra stop 'Haymarket' added to the Green Line).

NOTE: the appropriate json must be included at the top of mbta.js
or the grunt tests will break. Additionally, lines 90-95 must be
commented out for the grunt tests to work, but must be uncommented
for manual testing of the bonus problem.