Add 'mbta-no-comments.js' file
Exact copy of mbta.js without comments. The code printing out
detailed results is moved to a separate function, so that the
important functions are easier to read.

The code as is will load the original station map, and does
nothing on its own but will pass the grunt test. In order to
run manual tests, lines 101-106 must be uncommented. In order
to load the 'bonus' station map with 'Haymarket' added to the
beginning of the Green Line, 'bonus-stations.json' must be
required (line 4) instead of 'stations.json' (line 7).